Add license (CC0)
There's not too much here so it being in the public domain and not requiring people to credit me should be fine. I'd prefer credit if possible, but I don't want to require it so Mojang can use it in official Minecraft if they want to. Since I'm not redistributing any Minecraft assets without significant modification, I think I'm allowed to do this. I'm mostly just adding a licence so PRs don't cause a mess of copyright: https://choosealicense.com/no-permission/